requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.2.0
    # via -r requirements.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.111.1
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
passlib==1.7.4
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.31
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
uvicorn[standard]==0.30.4
    # via
    #   -r requirements.in
    #   fastapi
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
